Restructure for section library
