+++ docs/source/conf.py
@@ -16,7 +16,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('../../'))
 from sectionproperties import __version__ as ver
-autodoc_mock_imports = ['pybind11', 'meshpy', 'shapely', 'more_itertools', 'matplotlib', 'rhino_shapely_interop']
+autodoc_mock_imports = ['pybind11', 'meshpy', 'shapely', 'more_itertools', 'matplotlib', 'rhino_shapely_interop', 'mpl_toolkits']
 
 # -- Project information -----------------------------------------------------
