+++ sectionproperties/pre/library/nastran_sections.py
@@ -1,8 +1,8 @@
 import numpy as np
 from shapely.geometry import Polygon
-from sectionproperties.pre.sections import Geometry, CompoundGeometry
-from sectionproperties.pre import pre
-from sectionproperties.pre.sections import draw_radius
+from sectionproperties.pre.sections import Geometry
+import sectionproperties.pre.pre as pre
+from sectionproperties.pre.library.utils import draw_radius
 
 
 def nastran_bar(DIM1: float, DIM2: float, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
@@ -17,9 +17,9 @@ def nastran_bar(DIM1: float, DIM2: float, material: pre.Material = pre.DEFAULT_M
     The following example creates a BAR cross-section with a depth of 1.5 and width of 2.0, and
     generates a mesh with a maximum triangular area of 0.001::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_bar
 
-        geometry = nsections.nastran_bar(DIM1=2.0, DIM2=1.5)
+        geometry = nastran_bar(DIM1=2.0, DIM2=1.5)
         mesh = geometry.create_mesh(mesh_sizes=[0.001])
 
     ..  figure:: ../images/sections/bar_geometry.png
@@ -63,9 +63,9 @@ def nastran_box(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pr
     The following example creates a BOX cross-section with a depth of 3.0 and width of 4.0, and
     generates a mesh with a maximum triangular area of 0.001::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_box
 
-        geometry = nsections.nastran_box(DIM1=4.0, DIM2=3.0, DIM3=0.375, DIM4=0.5)
+        geometry = nastran_box(DIM1=4.0, DIM2=3.0, DIM3=0.375, DIM4=0.5)
         mesh = geometry.create_mesh(mesh_sizes=[0.001])
 
     ..  figure:: ../images/sections/box_geometry.png
@@ -83,23 +83,30 @@ def nastran_box(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pr
     # Ensure dimensions are physically relevant
     np.testing.assert_(2.0 * DIM4 < DIM1, "Invalid geometry specified.")
     np.testing.assert_(2.0 * DIM3 < DIM2, "Invalid geometry specified.")
-
-    points = [
+    points_outer = [
         [-0.5 * DIM1, -0.5 * DIM2],
         [0.5 * DIM1, -0.5 * DIM2],
         [0.5 * DIM1, 0.5 * DIM2],
         [-0.5 * DIM1, 0.5 * DIM2],
+    ]
+    points_inner = [
         [-0.5 * DIM1 + DIM4, -0.5 * DIM2 + DIM3],
         [0.5 * DIM1 - DIM4, -0.5 * DIM2 + DIM3],
         [0.5 * DIM1 - DIM4, 0.5 * DIM2 - DIM3],
         [-0.5 * DIM1 + DIM4, 0.5 * DIM2 - DIM3],
     ]
-    geometry = Geometry(Polygon(points), material)
+
+    inner_box = Polygon(points_inner)
+    outer_box = Polygon(points_outer)
+
     C = (0.5 * DIM1, 0.5 * DIM2)
     D = (0.5 * DIM1, -0.5 * DIM2)
     E = (-0.5 * DIM1, -0.5 * DIM2)
     F = (-0.5 * DIM1, 0.5 * DIM2)
+
+    geometry = Geometry(outer_box - inner_box, material)
     geometry.recovery_points = [C, D, E, F]
+
     return geometry
 
 
@@ -121,9 +128,9 @@ def nastran_box1(
     The following example creates a BOX1 cross-section with a depth of 3.0 and width of 4.0, and
     generates a mesh with a maximum triangular area of 0.007::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_box1
 
-        geometry = nsections.nastran_box1(
+        geometry = nastran_box1(
             DIM1=4.0, DIM2=3.0, DIM3=0.375, DIM4=0.5, DIM5=0.25, DIM6=0.75
         )
         mesh = geometry.create_mesh(mesh_sizes=[0.007])
@@ -180,9 +187,9 @@ def nastran_chan(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: p
     The following example creates a CHAN cross-section with a depth of 4.0 and width of 2.0, and
     generates a mesh with a maximum triangular area of 0.008::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_chan
 
-        geometry = nsections.nastran_chan(DIM1=2.0, DIM2=4.0, DIM3=0.25, DIM4=0.5)
+        geometry = nastran_chan(DIM1=2.0, DIM2=4.0, DIM3=0.25, DIM4=0.5)
         mesh = geometry.create_mesh(mesh_sizes=[0.008])
 
     ..  figure:: ../images/sections/chan_geometry.png
@@ -238,9 +245,9 @@ def nastran_chan1(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: 
     The following example creates a CHAN1 cross-section with a depth of 4.0 and width of 1.75, and
     generates a mesh with a maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_chan1
 
-        geometry = nsections.nastran_chan1(DIM1=0.75, DIM2=1.0, DIM3=3.5, DIM4=4.0)
+        geometry = nastran_chan1(DIM1=0.75, DIM2=1.0, DIM3=3.5, DIM4=4.0)
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
 
     ..  figure:: ../images/sections/chan1_geometry.png
@@ -294,9 +301,9 @@ def nastran_chan2(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: 
     The following example creates a CHAN2 cross-section with a depth of 2.0 and width of 4.0, and
     generates a mesh with a maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_chan2
 
-        geometry = nsections.nastran_chan2(DIM1=0.375, DIM2=0.5, DIM3=2.0, DIM4=4.0)
+        geometry = nastran_chan2(DIM1=0.375, DIM2=0.5, DIM3=2.0, DIM4=4.0)
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
 
     ..  figure:: ../images/sections/chan2_geometry.png
@@ -351,9 +358,9 @@ def nastran_cross(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: 
     The following example creates a rectangular cross-section with a depth of 3.0 and width of
     1.875, and generates a mesh with a maximum triangular area of 0.008::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_cross
 
-        geometry = nsections.nastran_cross(DIM1=1.5, DIM2=0.375, DIM3=3.0, DIM4=0.25)
+        geometry = nastran_cross(DIM1=1.5, DIM2=0.375, DIM3=3.0, DIM4=0.25)
         mesh = geometry.create_mesh(mesh_sizes=[0.008])
 
     ..  figure:: ../images/sections/cross_geometry.png
@@ -422,9 +429,9 @@ def nastran_fcross(
 
     The following example demonstrates the creation of a flanged cross section::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_fcross
 
-        geometry = nsections.nastran_fcross(
+        geometry = nastran_fcross(
             DIM1=9.0, DIM2=6.0, DIM3=0.75, DIM4=0.625, DIM5=2.1, DIM6=0.375, DIM7=4.5, DIM8=0.564
         )
         mesh = geometry.create_mesh(mesh_sizes=[0.03])
@@ -522,9 +529,9 @@ def nastran_dbox(
     The following example creates a DBOX cross-section with a depth of 3.0 and width of 8.0, and
     generates a mesh with a maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_dbox
 
-        geometry = nsections.nastran_dbox(
+        geometry = nastran_dbox(
             DIM1=8.0, DIM2=3.0, DIM3=3.0, DIM4=0.5, DIM5=0.625, DIM6=0.75, DIM7=0.375, DIM8=0.25,
             DIM9=0.5, DIM10=0.375
         )
@@ -597,9 +604,9 @@ def nastran_gbox(
     The following example creates a GBOX cross-section with a depth of 2.5 and width of 6.0, and
     generates a mesh with a maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_gbox
 
-        geometry = nsections.nastran_gbox(
+        geometry = nastran_gbox(
             DIM1=6.0, DIM2=2.5, DIM3=0.375, DIM4=0.25, DIM5=0.625, DIM6=1.0
         )
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
@@ -667,9 +674,9 @@ def nastran_h(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pre.
     The following example creates a H cross-section with a depth of 3.5 and width of 2.75, and
     generates a mesh with a maximum triangular area of 0.005::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_h
 
-        geometry = nsections.nastran_h(DIM1=2.0, DIM2=0.75, DIM3=3.5, DIM4=0.25)
+        geometry = nastran_h(DIM1=2.0, DIM2=0.75, DIM3=3.5, DIM4=0.25)
         mesh = geometry.create_mesh(mesh_sizes=[0.005])
 
     ..  figure:: ../images/sections/h_geometry.png
@@ -728,9 +735,9 @@ def nastran_hat(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pr
     The following example creates a HAT cross-section with a depth of 1.25 and width of 2.5, and
     generates a mesh with a maximum triangular area of 0.001::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_hat
 
-        geometry = nsections.nastran_hat(DIM1=1.25, DIM2=0.25, DIM3=1.5, DIM4=0.5)
+        geometry = nastran_hat(DIM1=1.25, DIM2=0.25, DIM3=1.5, DIM4=0.5)
         mesh = geometry.create_mesh(mesh_sizes=[0.001])
 
     ..  figure:: ../images/sections/hat_geometry.png
@@ -789,9 +796,9 @@ def nastran_hat1(DIM1: float, DIM2: float, DIM3: float, DIM4: float, DIM5: float
     The following example creates a HAT1 cross-section with a depth of 2.0 and width of 4.0, and
     generates a mesh with a maximum triangular area of 0.005::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_hat1
 
-        geometry = nsections.nastran_hat1(DIM1=4.0, DIM2=2.0, DIM3=1.5, DIM4=0.1875, DIM5=0.375)
+        geometry = nastran_hat1(DIM1=4.0, DIM2=2.0, DIM3=1.5, DIM4=0.1875, DIM5=0.375)
         mesh = geometry.create_mesh(mesh_sizes=[0.005])
 
     ..  figure:: ../images/sections/hat1_geometry.png
@@ -847,9 +854,9 @@ def nastran_hexa(DIM1: float, DIM2: float, DIM3: float, material: pre.Material =
     The following example creates a rectangular cross-section with a depth of 1.5 and width of 2.0,
     and generates a mesh with a maximum triangular area of 0.005::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_hexa
 
-        geometry = nsections.nastran_hexa(DIM1=0.5, DIM2=2.0, DIM3=1.5)
+        geometry = nastran_hexa(DIM1=0.5, DIM2=2.0, DIM3=1.5)
         mesh = geometry.create_mesh(mesh_sizes=[0.005])
 
     ..  figure:: ../images/sections/hexa_geometry.png
@@ -905,9 +912,9 @@ def nastran_i(
     The following example creates a Nastran I cross-section with a depth of 5.0, and generates a
     mesh with a maximum triangular area of 0.008::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_i
 
-        geometry = nsections.nastran_i(
+        geometry = nastran_i(
             DIM1=5.0, DIM2=2.0, DIM3=3.0, DIM4=0.25, DIM5=0.375, DIM6=0.5
         )
         mesh = geometry.create_mesh(mesh_sizes=[0.008])
@@ -972,9 +979,9 @@ def nastran_i1(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pre
     5.0 and width of 1.75, and generates a mesh with a maximum triangular area of
     0.02::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_i1
 
-        geometry = nsections.nastran_i1(DIM1=1.0, DIM2=0.75, DIM3=4.0, DIM4=5.0)
+        geometry = nastran_i1(DIM1=1.0, DIM2=0.75, DIM3=4.0, DIM4=5.0)
         mesh = geometry.create_mesh(mesh_sizes=[0.02])
 
     ..  figure:: ../images/sections/i1_geometry.png
@@ -1033,9 +1040,9 @@ def nastran_l(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pre.
     The following example creates a L cross-section with a depth of 6.0 and width of 3.0, and
     generates a mesh with a maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_l
 
-        geometry = nsections.nastran_l(DIM1=3.0, DIM2=6.0, DIM3=0.375, DIM4=0.625)
+        geometry = nastran_l(DIM1=3.0, DIM2=6.0, DIM3=0.375, DIM4=0.625)
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
 
     ..  figure:: ../images/sections/l_geometry.png
@@ -1064,7 +1071,7 @@ def nastran_l(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pre.
     E = (-0.5 * DIM4, -0.5 * DIM3)
     F = (-0.5 * DIM4, DIM2 - 0.5 * DIM3)
 
-    geometry.recovery_points([C, D, E, F])
+    geometry.recovery_points = [C, D, E, F]
     return geometry
 
 
@@ -1080,9 +1087,9 @@ def nastran_rod(DIM1: float, n: int, material: pre.Material = pre.DEFAULT_MATERI
     The following example creates a circular rod with a radius of 3.0 and 50 points discretising
     the boundary, and generates a mesh with a maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_rod
 
-        geometry = nsections.nastran_rod(DIM1=3.0, n=50)
+        geometry = nastran_rod(DIM1=3.0, n=50)
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
 
     ..  figure:: ../images/sections/rod_geometry.png
@@ -1118,7 +1125,7 @@ def nastran_rod(DIM1: float, n: int, material: pre.Material = pre.DEFAULT_MATERI
     E = (0, -DIM1)
     F = (-DIM1, 0)
 
-    geometry.recovery_points([C, D, E, F])
+    geometry.recovery_points = [C, D, E, F]
 
     return geometry
 
@@ -1137,9 +1144,9 @@ def nastran_tee(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pr
     The following example creates a T cross-section with a depth of 4.0 and width of 3.0, and
     generates a mesh with a maximum triangular area of 0.001::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_tee
 
-        geometry = nsections.nastran_t(DIM1=3.0, DIM2=4.0, DIM3=0.375, DIM4=0.25)
+        geometry = nastran_tee(DIM1=3.0, DIM2=4.0, DIM3=0.375, DIM4=0.25)
         mesh = geometry.create_mesh(mesh_sizes=[0.001])
 
     ..  figure:: ../images/sections/t_geometry.png
@@ -1210,9 +1217,9 @@ def nastran_tee1(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: p
     The following example creates a T1 cross-section with a depth of 3.0 and width of 3.875, and
     generates a mesh with a maximum triangular area of 0.001::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_tee1
 
-        geometry = nsections.nastran_t1(DIM1=3.0, DIM2=3.5, DIM3=0.375, DIM4=0.25)
+        geometry = nastran_tee1(DIM1=3.0, DIM2=3.5, DIM3=0.375, DIM4=0.25)
         mesh = geometry.create_mesh(mesh_sizes=[0.001])
 
     ..  figure:: ../images/sections/t1_geometry.png
@@ -1268,9 +1275,9 @@ def nastran_tee2(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: p
     The following example creates a T2 cross-section with a depth of 4.0 and width of 3.0, and
     generates a mesh with a maximum triangular area of 0.005::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_tee2
 
-        geometry = nsections.nastran_t2(DIM1=3.0, DIM2=4.0, DIM3=0.375, DIM4=0.5)
+        geometry = nastran_tee2(DIM1=3.0, DIM2=4.0, DIM3=0.375, DIM4=0.5)
         mesh = geometry.create_mesh(mesh_sizes=[0.005])
 
     ..  figure:: ../images/sections/t2_geometry.png
@@ -1326,9 +1333,9 @@ def nastran_tube(DIM1: float, DIM2: float, n: int, material: pre.Material = pre.
     inner radius of 2.5, and generates a mesh with 37 points discretising the boundaries and a
     maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_tube
 
-        geometry = nsections.nastran_tube(DIM1=3.0, DIM2=2.5, n=37)
+        geometry = nastran_tube(DIM1=3.0, DIM2=2.5, n=37)
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
 
     ..  figure:: ../images/sections/tube_geometry.png
@@ -1392,9 +1399,9 @@ def nastran_tube2(DIM1: float, DIM2: float, n: float, material: pre.Material = p
     wall thickness of 0.5, and generates a mesh with 37 point discretising the boundary and a
     maximum triangular area of 0.01::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_tube2
 
-        geometry = nsections.nastran_tube2(DIM1=3.0, DIM2=0.5, n=37)
+        geometry = nastran_tube2(DIM1=3.0, DIM2=0.5, n=37)
         mesh = geometry.create_mesh(mesh_sizes=[0.01])
 
     ..  figure:: ../images/sections/tube2_geometry.png
@@ -1433,16 +1440,16 @@ def nastran_tube2(DIM1: float, DIM2: float, n: float, material: pre.Material = p
         points_outer.append([x_outer, y_outer])
         points_inner.append([x_inner, y_inner])
 
-        exterior = Geometry(Polygon(points_outer), material)
-        interior = Geometry(Polygon(points_inner), material)
-        geometry = exterior - interior
+    exterior = Geometry(Polygon(points_outer), material)
+    interior = Geometry(Polygon(points_inner), material)
+    geometry = exterior - interior
 
-        C = (0, DIM1)
-        D = (DIM1, 0)
-        E = (0, -DIM1)
-        F = (-DIM1, 0)
-        geometry.recovery_points = [C, D, E, F]
-        return geometry
+    C = (0, DIM1)
+    D = (DIM1, 0)
+    E = (0, -DIM1)
+    F = (-DIM1, 0)
+    geometry.recovery_points = [C, D, E, F]
+    return geometry
 
 
 def nastran_zed(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
@@ -1459,9 +1466,9 @@ def nastran_zed(DIM1: float, DIM2: float, DIM3: float, DIM4: float, material: pr
     The following example creates a rectangular cross-section with a depth of 4.0 and width of
     2.75, and generates a mesh with a maximum triangular area of 0.005::
 
-        import sectionproperties.pre.nastran_sections as nsections
+        from sectionproperties.pre.library.nastran_sections import nastran_zed
 
-        geometry = nsections.nastran_z(DIM1=1.125, DIM2=0.5, DIM3=3.5, DIM4=4.0)
+        geometry = nastran_zed(DIM1=1.125, DIM2=0.5, DIM3=3.5, DIM4=4.0)
         mesh = geometry.create_mesh(mesh_sizes=[0.005])
 
     ..  figure:: ../images/sections/z_geometry.png
