+++ sectionproperties/pre/sections.py
@@ -16,7 +16,6 @@ from shapely.ops import split, unary_union
 import shapely
 import matplotlib
 import matplotlib.pyplot as plt
-# from IPython.display import display_svg
 import sectionproperties.pre.pre as pre
 import sectionproperties.pre.bisect_section as bisect
 import sectionproperties.post.post as post
@@ -209,16 +208,16 @@ class Geometry:
     def from_3dm(cls, filepath: Union[str, pathlib.Path], **kwargs) -> Geometry:
         """Class method to create a `Geometry` from the objects in a Rhino `.3dm` file.
 
-        :param filepath: 
+        :param filepath:
             File path to the rhino `.3dm` file.
         :type filepath: Union[str, pathlib.Path]
         :param \**kwargs:
             See below.
-        :raises RuntimeError: 
-            A RuntimeError is raised if two or more polygons are found. 
-            This is dependent on the keyword arguments. 
-            Try adjusting the keyword arguments if this error is raised. 
-        :return: 
+        :raises RuntimeError:
+            A RuntimeError is raised if two or more polygons are found.
+            This is dependent on the keyword arguments.
+            Try adjusting the keyword arguments if this error is raised.
+        :return:
             A Geometry object.
         :rtype: :class:`sectionproperties.pre.sections.Geometry`
 
@@ -226,7 +225,7 @@ class Geometry:
             * *refine_num* (``int, optional``) --
                 Bézier curve interpolation number. In Rhino a surface's edges are nurb based curves.
                 Shapely does not support nurbs, so the individual Bézier curves are interpolated using straight lines.
-                This parameter sets the number of straight lines used in the interpolation. 
+                This parameter sets the number of straight lines used in the interpolation.
                 Default is 1.
             * *vec1* (``numpy.ndarray, optional``) --
                 A 3d vector in the Shapely plane. Rhino is a 3D geometry environment.
@@ -264,12 +263,12 @@ class Geometry:
     def from_rhino_encoding(cls, r3dm_brep: str, **kwargs) -> Geometry:
         """Load an encoded single surface planer brep.
 
-        :param r3dm_brep: 
+        :param r3dm_brep:
             A Rhino3dm.Brep encoded as a string.
         :type r3dm_brep: str
         :param \**kwargs:
             See below.
-        :return: 
+        :return:
             A Geometry object found in the encoded string.
         :rtype: :class:`sectionproperties.pre.sections.Geometry`
 
@@ -277,7 +276,7 @@ class Geometry:
             * *refine_num* (``int, optional``) --
                 Bézier curve interpolation number. In Rhino a surface's edges are nurb based curves.
                 Shapely does not support nurbs, so the individual Bézier curves are interpolated using straight lines.
-                This parameter sets the number of straight lines used in the interpolation. 
+                This parameter sets the number of straight lines used in the interpolation.
                 Default is 1.
             * *vec1* (``numpy.ndarray, optional``) --
                 A 3d vector in the Shapely plane. Rhino is a 3D geometry environment.
@@ -299,7 +298,7 @@ class Geometry:
                 Controls if only the rhino surfaces that have the same normal as the Shapely plane are yielded.
                 If true, all non parallel surfaces are filtered out.
                 Default is False.
-        """ 
+        """
         return cls(geom = rhino_importer.load_brep_encoding(r3dm_brep, **kwargs)[0])
 
     def create_facets_and_control_points(self):
@@ -967,7 +966,7 @@ class Geometry:
         try:
             new_polygon = self.geom | other.geom
             if isinstance(new_polygon, MultiPolygon):
-                
+
                 return CompoundGeometry(
                     [Geometry(polygon, material) for polygon in new_polygon.geoms]
                 )
@@ -1041,7 +1040,7 @@ class Geometry:
             raise ValueError(
                 f"Cannot perform 'intersection' on these two Geometry instances: {self} & {other}"
             )
-###
+
 
 class CompoundGeometry(Geometry):
     """Class for defining a geometry of multiple distinct regions, each potentially
@@ -1155,7 +1154,7 @@ class CompoundGeometry(Geometry):
                 continue
             prev_j_idx = prev_facet[1]
             if i_idx != prev_j_idx:  # If there is a break in the chain of edges...
-                # ... then add the last point, close off the polygon, 
+                # ... then add the last point, close off the polygon,
                 # and add the polygon to the all_polygons accumulator....
                 current_polygon_points.append(points[prev_j_idx])
                 all_polygons.append(Polygon(current_polygon_points))
@@ -1217,12 +1216,12 @@ class CompoundGeometry(Geometry):
     def from_3dm(cls, filepath: Union[str, pathlib.Path], **kwargs) -> CompoundGeometry:
         """Class method to create a `CompoundGeometry` from the objects in a Rhino `3dm` file.
 
-        :param filepath: 
+        :param filepath:
             File path to the rhino `.3dm` file.
         :type filepath: Union[str, pathlib.Path]
         :param \**kwargs:
             See below.
-        :return: 
+        :return:
             A `CompoundGeometry` object.
         :rtype: :class:`sectionproperties.pre.sections.CompoundGeometry`
 
@@ -1230,7 +1229,7 @@ class CompoundGeometry(Geometry):
             * *refine_num* (``int, optional``) --
                 Bézier curve interpolation number. In Rhino a surface's edges are nurb based curves.
                 Shapely does not support nurbs, so the individual Bézier curves are interpolated using straight lines.
-                This parameter sets the number of straight lines used in the interpolation. 
+                This parameter sets the number of straight lines used in the interpolation.
                 Default is 1.
             * *vec1* (``numpy.ndarray, optional``) --
                 A 3d vector in the Shapely plane. Rhino is a 3D geometry environment.
@@ -1252,7 +1251,7 @@ class CompoundGeometry(Geometry):
                 Controls if only the rhino surfaces that have the same normal as the Shapely plane are yielded.
                 If true, all non parallel surfaces are filtered out.
                 Default is False.
-        """  
+        """
         list_poly = rhino_importer.load_3dm(filepath, **kwargs)
         return cls(geoms = MultiPolygon(list_poly))
 
@@ -1684,1412 +1683,3 @@ def create_points_and_facets(shape: Polygon, tol=12) -> tuple:
         points += int_points
 
     return points, facets
-
-
-def draw_radius(
-    pt: list, r: float, theta: float, n, ccw: bool = True
-):  # Changed 'anti' to ccw to match shapely
-    """Adds a quarter radius of points to the points list - centered at point *pt*, with radius
-    *r*, starting at angle *theta*, with *n* points. If r = 0, adds pt only.
-
-    :param pt: Centre of radius *(x,y)*
-    :type pt: list[float, float]
-    :param float r: Radius
-    :param float theta: Initial angle
-    :param int n: Number of points
-    :param bool ccw: Counter-clockwise rotation?
-    """
-    points = []
-    if r == 0:
-        points.append(pt)
-        return points
-
-    if ccw:
-        mult = 1
-    else:
-        mult = -1
-
-    # calculate radius of points
-    for i in range(n):
-        # determine angle
-        t = theta + mult * i * 1.0 / max(1, n - 1) * np.pi * 0.5
-
-        x = pt[0] + r * np.cos(t)
-        y = pt[1] + r * np.sin(t)
-        points.append([x, y])
-    return points
-
-
-def rectangular_section(b, d, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a rectangular section with the bottom left corner at the origin *(0, 0)*, with
-    depth *d* and width *b*.
-
-    :param float d: Depth (y) of the rectangle
-    :param float b: Width (x) of the rectangle
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a rectangular cross-section with a depth of 100 and width of 50,
-    and generates a mesh with a maximum triangular area of 5::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.rectangular_section(d=100, b=50)
-        geometry.create_mesh(mesh_sizes=[5])
-
-    ..  figure:: ../images/sections/rectangle_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Rectangular section geometry.
-
-    ..  figure:: ../images/sections/rectangle_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = [[0, 0], [b, 0], [b, d], [0, d]]
-    rectangle = Polygon(points)
-    return Geometry(rectangle, material)
-
-
-def circular_section(d: float, n: int, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a solid circle centered at the origin *(0, 0)* with diameter *d* and using *n*
-    points to construct the circle.
-
-    :param float d: Diameter of the circle
-    :param int n: Number of points discretising the circle
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a circular geometry with a diameter of 50 with 64 points,
-    and generates a mesh with a maximum triangular area of 2.5::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.circular_section(d=50, n=64)
-        geometry.create_mesh(mesh_sizes=[2.5])
-
-    ..  figure:: ../images/sections/circle_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Circular section geometry.
-
-    ..  figure:: ../images/sections/circle_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    x_off, y_off = (0, 0)
-    points = []
-    # loop through each point on the circle
-    for i in range(n):
-        # determine polar angle
-        theta = i * 2 * np.pi * 1.0 / n
-
-        # calculate location of the point
-        x = 0.5 * d * np.cos(theta) + x_off
-        y = 0.5 * d * np.sin(theta) + y_off
-
-        # append the current point to the points list
-        points.append([x, y])
-
-    circle = Polygon(points)
-    return Geometry(circle, material)
-
-
-def circular_hollow_section(d: float, t: float, n: int, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a circular hollow section (CHS) centered at the origin *(0, 0)*, with diameter *d* and
-    thickness *t*, using *n* points to construct the inner and outer circles.
-
-    :param float d: Outer diameter of the CHS
-    :param float t: Thickness of the CHS
-    :param int n: Number of points discretising the inner and outer circles
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a CHS discretised with 64 points, with a diameter of 48 and
-    thickness of 3.2, and generates a mesh with a maximum triangular area of 1.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.circular_hollow_section(d=48, t=3.2, n=64)
-        geometry.create_mesh(mesh_sizes=[1.0])
-
-    ..  figure:: ../images/sections/chs_geometry.png
-        :align: center
-        :scale: 75 %
-
-        CHS geometry.
-
-    ..  figure:: ../images/sections/chs_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points_inner = []
-    points_outer = []
-    # loop through each point of the CHS
-    for i in range(n):
-        # determine polar angle
-        theta = i * 2 * np.pi * 1.0 / n
-
-        # calculate location of outer and inner points
-        x_outer = 0.5 * d * np.cos(theta)
-        y_outer = 0.5 * d * np.sin(theta)
-        x_inner = (0.5 * d - t) * np.cos(theta)
-        y_inner = (0.5 * d - t) * np.sin(theta)
-
-        # append the current points to the points list
-        points_outer.append([x_outer, y_outer])
-        points_inner.append([x_inner, y_inner])
-
-    inner_circle = Polygon(points_inner)
-    outer_circle = Polygon(points_outer)
-    return Geometry(outer_circle - inner_circle, material)
-
-
-def elliptical_section(d_y: float, d_x: float, n: int, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a solid ellipse centered at the origin *(0, 0)* with vertical diameter *d_y* and
-    horizontal diameter *d_x*, using *n* points to construct the ellipse.
-
-    :param float d_y: Diameter of the ellipse in the y-dimension
-    :param float d_x: Diameter of the ellipse in the x-dimension
-    :param int n: Number of points discretising the ellipse
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates an elliptical cross-section with a vertical diameter of 25 and
-    horizontal diameter of 50, with 40 points, and generates a mesh with a maximum triangular area
-    of 1.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.elliptical_section(d_y=25, d_x=50, n=40)
-        geometry.create_mesh(mesh_sizes=[1.0])
-
-    ..  figure:: ../images/sections/ellipse_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Elliptical section geometry.
-
-    ..  figure:: ../images/sections/ellipse_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-
-    # loop through each point on the ellipse
-    for i in range(n):
-        # determine polar angle
-        theta = i * 2 * np.pi * 1.0 / n
-
-        # calculate location of the point
-        x = 0.5 * d_x * np.cos(theta)
-        y = 0.5 * d_y * np.sin(theta)
-
-        # append the current point to the points list
-        points.append([x, y])
-
-    ellipse = Polygon(points)
-    return Geometry(ellipse, material)
-
-
-def elliptical_hollow_section(d_y: float, d_x: float, t: float, n: int, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs an elliptical hollow section (EHS) centered at the origin *(0, 0)*, with outer vertical
-    diameter *d_y*, outer horizontal diameter *d_x*, and thickness *t*, using *n* points to
-    construct the inner and outer ellipses.
-
-    :param float d_y: Diameter of the ellipse in the y-dimension
-    :param float d_x: Diameter of the ellipse in the x-dimension
-    :param float t: Thickness of the EHS
-    :param int n: Number of points discretising the inner and outer ellipses
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a EHS discretised with 30 points, with a outer vertical diameter
-    of 25, outer horizontal diameter of 50, and thickness of 2.0, and generates a mesh with a
-    maximum triangular area of 0.5::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.elliptical_hollow_section(d_y=25, d_x=50, t=2.0, n=64)
-        geometry.create_mesh(mesh_sizes=[0.5])
-
-    ..  figure:: ../images/sections/ehs_geometry.png
-        :align: center
-        :scale: 75 %
-
-        EHS geometry.
-
-    ..  figure:: ../images/sections/ehs_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points_inner = []
-    points_outer = []
-    # loop through each point of the EHS
-    for i in range(n):
-        # determine polar angle
-        theta = i * 2 * np.pi * 1.0 / n
-
-        # calculate location of outer and inner points
-        x_outer = 0.5 * d_x * np.cos(theta)
-        y_outer = 0.5 * d_y * np.sin(theta)
-        x_inner = ((0.5 * d_x) - t) * np.cos(theta)
-        y_inner = ((0.5 * d_y) - t) * np.sin(theta)
-
-        # append the current points to the points list
-        points_outer.append([x_outer, y_outer])
-        points_inner.append([x_inner, y_inner])
-
-    outer = Polygon(points_outer)
-    inner = Polygon(points_inner)
-    return Geometry(outer - inner, material)
-
-
-def rectangular_hollow_section(b: float, d: float, t: float, r_out: float, n_r: int, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a rectangular hollow section (RHS) centered at *(b/2, d/2)*, with depth *d*, width *b*,
-    thickness *t* and outer radius *r_out*, using *n_r* points to construct the inner and outer
-    radii. If the outer radius is less than the thickness of the RHS, the inner radius is set to
-    zero.
-
-    :param float d: Depth of the RHS
-    :param float b: Width of the RHS
-    :param float t: Thickness of the RHS
-    :param float r_out: Outer radius of the RHS
-    :param int n_r: Number of points discretising the inner and outer radii
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates an RHS with a depth of 100, a width of 50, a thickness of 6 and
-    an outer radius of 9, using 8 points to discretise the inner and outer radii. A mesh is
-    generated with a maximum triangular area of 2.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.rectangular_hollow_section(d=100, b=50, t=6, r_out=9, n_r=8)
-        geometry.create_mesh(mesh_sizes=[2.0])
-
-    ..  figure:: ../images/sections/rhs_geometry.png
-        :align: center
-        :scale: 75 %
-
-        RHS geometry.
-
-    ..  figure:: ../images/sections/rhs_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points_inner = []
-    points_outer = []
-    # calculate internal radius
-    r_in = max(r_out - t, 0)
-    # construct the outer radius points
-    points_outer += draw_radius([r_out, r_out], r_out, np.pi, n_r)
-    points_outer += draw_radius([b - r_out, r_out], r_out, 1.5 * np.pi, n_r)
-    points_outer += draw_radius([b - r_out, d - r_out], r_out, 0, n_r)
-    points_outer += draw_radius([r_out, d - r_out], r_out, 0.5 * np.pi, n_r)
-
-    points_inner += draw_radius([t + r_in, t + r_in], r_in, np.pi, n_r)
-    points_inner += draw_radius([b - t - r_in, t + r_in], r_in, 1.5 * np.pi, n_r)
-    points_inner += draw_radius([b - t - r_in, d - t - r_in], r_in, 0, n_r)
-    points_inner += draw_radius([t + r_in, d - t - r_in], r_in, 0.5 * np.pi, n_r)
-
-    outer = Polygon(points_outer)
-    inner = Polygon(points_inner)
-    return Geometry(outer - inner, material)
-
-
-def i_section(
-    d: float, b: float, t_f: float, t_w: float, r: float, n_r: int, material: pre.Material = pre.DEFAULT_MATERIAL
-) -> Geometry:  # More specific description and less ambiguous? e.g. not an "S" section.
-    """Constructs an I-section centered at *(b/2, d/2)*, with depth *d*, width *b*, flange
-    thickness *t_f*, web thickness *t_w*, and root radius *r*, using *n_r* points to construct the
-    root radius.
-
-    :param float d: Depth of the I-section
-    :param float b: Width of the I-section
-    :param float t_f: Flange thickness of the I-section
-    :param float t_w: Web thickness of the I-section
-    :param float r: Root radius of the I-section
-    :param int n_r: Number of points discretising the root radius
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates an I-section with a depth of 203, a width of 133, a flange
-    thickness of 7.8, a web thickness of 5.8 and a root radius of 8.9, using 16 points to
-    discretise the root radius. A mesh is generated with a maximum triangular area of 3.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.i_section(d=203, b=133, t_f=7.8, t_w=5.8, r=8.9, n_r=16)
-        geometry.create_mesh(mesh_sizes=[3.0])
-
-    ..  figure:: ../images/sections/isection_geometry.png
-        :align: center
-        :scale: 75 %
-
-        I-section geometry.
-
-    ..  figure:: ../images/sections/isection_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-
-    # add first three points
-    points.append([0, 0])
-    points.append([b, 0])
-    points.append([b, t_f])
-
-    # construct the bottom right radius
-    pt = [b * 0.5 + t_w * 0.5 + r, t_f + r]
-    points += draw_radius(pt, r, 1.5 * np.pi, n_r, False)
-
-    # construct the top right radius
-    pt = [b * 0.5 + t_w * 0.5 + r, d - t_f - r]
-    points += draw_radius(pt, r, np.pi, n_r, False)
-
-    # add the next four points
-    points.append([b, d - t_f])
-    points.append([b, d])
-    points.append([0, d])
-    points.append([0, d - t_f])
-
-    # construct the top left radius
-    pt = [b * 0.5 - t_w * 0.5 - r, d - t_f - r]
-    points += draw_radius(pt, r, 0.5 * np.pi, n_r, False)
-
-    # construct the bottom left radius
-    pt = [b * 0.5 - t_w * 0.5 - r, t_f + r]
-    points += draw_radius(pt, r, 0, n_r, False)
-
-    # # add the last point
-    points.append([0, t_f])
-    i_section = Polygon(points)
-    return Geometry(i_section, material)
-
-
-def mono_i_section(d, b_t, b_b, t_fb, t_ft, t_w, r, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a monosymmetric I-section centered at *(max(b_t, b_b)/2, d/2)*, with depth *d*,
-    top flange width *b_t*, bottom flange width *b_b*, top flange thickness *t_ft*, top flange
-    thickness *t_fb*, web thickness *t_w*, and root radius *r*, using *n_r* points to construct the
-    root radius.
-
-    :param float d: Depth of the I-section
-    :param float b_t: Top flange width
-    :param float b_b: Bottom flange width
-    :param float t_ft: Top flange thickness of the I-section
-    :param float t_fb: Bottom flange thickness of the I-section
-    :param float t_w: Web thickness of the I-section
-    :param float r: Root radius of the I-section
-    :param int n_r: Number of points discretising the root radius
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a monosymmetric I-section with a depth of 200, a top flange width
-    of 50, a top flange thickness of 12, a bottom flange width of 130, a bottom flange thickness of
-    8, a web thickness of 6 and a root radius of 8, using 16 points to discretise the root radius.
-    A mesh is generated with a maximum triangular area of 3.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.mono_i_section(
-            d=200, b_t=50, b_b=130, t_ft=12, t_fb=8, t_w=6, r=8, n_r=16
-        )
-        geometry.create_mesh(mesh_sizes=[3.0])
-
-    ..  figure:: ../images/sections/monoisection_geometry.png
-        :align: center
-        :scale: 75 %
-
-        I-section geometry.
-
-    ..  figure:: ../images/sections/monoisection_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-    # calculate central axis
-    x_central = max(b_t, b_b) * 0.5
-
-    # add first three points
-    points.append([x_central - b_b * 0.5, 0])
-    points.append([x_central + b_b * 0.5, 0])
-    points.append([x_central + b_b * 0.5, t_fb])
-
-    # construct the bottom right radius
-    pt = [x_central + t_w * 0.5 + r, t_fb + r]
-    points += draw_radius(pt, r, 1.5 * np.pi, n_r, False)
-
-    # construct the top right radius
-    pt = [x_central + t_w * 0.5 + r, d - t_ft - r]
-    points += draw_radius(pt, r, np.pi, n_r, False)
-
-    # add the next four points
-    points.append([x_central + b_t * 0.5, d - t_ft])
-    points.append([x_central + b_t * 0.5, d])
-    points.append([x_central - b_t * 0.5, d])
-    points.append([x_central - b_t * 0.5, d - t_ft])
-
-    # construct the top left radius
-    pt = [x_central - t_w * 0.5 - r, d - t_ft - r]
-    points += draw_radius(pt, r, 0.5 * np.pi, n_r, False)
-
-    # construct the bottom left radius
-    pt = [x_central - t_w * 0.5 - r, t_fb + r]
-    points += draw_radius(pt, r, 0, n_r, False)
-
-    # add the last point
-    points.append([x_central - b_b * 0.5, t_fb])
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def tapered_flange_i_section(d, b, t_f, t_w, r_r, r_f, alpha, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a Tapered Flange I-section centered at *(b/2, d/2)*, with depth *d*, width *b*,
-    mid-flange thickness *t_f*, web thickness *t_w*, root radius *r_r*, flange radius *r_f* and
-    flange angle *alpha*, using *n_r* points to construct the radii.
-
-    :param float d: Depth of the Tapered Flange I-section
-    :param float b: Width of the Tapered Flange I-section
-    :param float t_f: Mid-flange thickness of the Tapered Flange I-section (measured at the point
-        equidistant from the face of the web to the edge of the flange)
-    :param float t_w: Web thickness of the Tapered Flange I-section
-    :param float r_r: Root radius of the Tapered Flange I-section
-    :param float r_f: Flange radius of the Tapered Flange I-section
-    :param float alpha: Flange angle of the Tapered Flange I-section (degrees)
-    :param int n_r: Number of points discretising the radii
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a Tapered Flange I-section with a depth of 588, a width of 191, a
-    mid-flange thickness of 27.2, a web thickness of 15.2, a root radius of 17.8, a flange radius
-    of 8.9 and a flange angle of 8°, using 16 points to discretise the radii. A mesh is generated
-    with a maximum triangular area of 20.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.tapered_flange_i_section(
-            d=588, b=191, t_f=27.2, t_w=15.2, r_r=17.8, r_f=8.9, alpha=8, n_r=16
-        )
-        geometry.create_mesh(mesh_sizes=[20.0])
-
-    ..  figure:: ../images/sections/taperedisection_geometry.png
-        :align: center
-        :scale: 75 %
-
-        I-section geometry.
-
-    ..  figure:: ../images/sections/taperedisection_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-
-    # calculate alpha in radians
-    alpha_rad = np.pi * alpha / 180
-
-    # calculate the height of the flange toe and dimensions of the straight
-    x1 = b * 0.25 - t_w * 0.25 - r_f * (1 - np.sin(alpha_rad))
-    y1 = x1 * np.tan(alpha_rad)
-    x2 = b * 0.25 - t_w * 0.25 - r_r * (1 - np.sin(alpha_rad))
-    y2 = x2 * np.tan(alpha_rad)
-    y_t = t_f - y1 - r_f * np.cos(alpha_rad)
-
-    # add first two points
-    points.append([0, 0])
-    points.append([b, 0])
-
-    # construct the bottom right flange toe radius
-    if r_f == 0:
-        points.append([b, y_t])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-
-            # calculate the locations of the radius points
-            x = b - r_f + r_f * np.cos(theta)
-            y = y_t + r_f * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the bottom right root radius
-    if r_r == 0:
-        points.append([b * 0.5 + t_w * 0.5, t_f + y2])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = (3.0 / 2 * np.pi - alpha_rad) - (
-                i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-            )
-
-            # calculate the locations of the radius points
-            x = b * 0.5 + t_w * 0.5 + r_r + r_r * np.cos(theta)
-            y = t_f + y2 + r_r * np.cos(alpha_rad) + r_r * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the top right root radius
-    if r_r == 0:
-        points.append([b * 0.5 + t_w * 0.5, d - t_f - y2])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = np.pi - i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-
-            # calculate the locations of the radius points
-            x = b * 0.5 + t_w * 0.5 + r_r + r_r * np.cos(theta)
-            y = d - t_f - y2 - r_r * np.cos(alpha_rad) + r_r * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the top right flange toe radius
-    if r_f == 0:
-        points.append([b, d - y_t])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = (3.0 * np.pi / 2 + alpha_rad) + i * 1.0 / max(1, n_r - 1) * (
-                np.pi * 0.5 - alpha_rad
-            )
-
-            # calculate the locations of the radius points
-            x = b - r_f + r_f * np.cos(theta)
-            y = d - y_t + r_f * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # add the next two points
-    points.append([b, d])
-    points.append([0, d])
-
-    # construct the top left flange toe radius
-    if r_f == 0:
-        points.append([0, d - y_t])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = np.pi + (i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad))
-
-            # calculate the locations of the radius points
-            x = r_f + r_f * np.cos(theta)
-            y = d - y_t + r_f * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the top left root radius
-    if r_r == 0:
-        points.append([b * 0.5 - t_w * 0.5, d - t_f - y2])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = (np.pi * 0.5 - alpha_rad) - (
-                i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-            )
-
-            # calculate the locations of the radius points
-            x = b * 0.5 - t_w * 0.5 - r_r + r_r * np.cos(theta)
-            y = d - t_f - y2 - r_r * np.cos(alpha_rad) + r_r * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the bottom left root radius
-    if r_r == 0:
-        points.append([b * 0.5 - t_w * 0.5, t_f + y2])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = -i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-
-            # calculate the locations of the radius points
-            x = b * 0.5 - t_w * 0.5 - r_r + r_r * np.cos(theta)
-            y = t_f + y2 + r_r * np.cos(alpha_rad) + r_r * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the bottom left flange toe radius
-    if r_f == 0:
-        points.append([0, y_t])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = (np.pi * 0.5 + alpha_rad) + (
-                i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-            )
-
-            # calculate the locations of the radius points
-            x = r_f + r_f * np.cos(theta)
-            y = y_t + r_f * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def channel_section(d, b, t_f, t_w, r, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a parallel-flange channel (PFC) section with the bottom left corner at the origin *(0, 0)*, with depth *d*,
-    width *b*, flange thickness *t_f*, web  thickness *t_w* and root radius *r*, using *n_r* points
-    to construct the root radius.
-
-    :param float d: Depth of the PFC section
-    :param float b: Width of the PFC section
-    :param float t_f: Flange thickness of the PFC section
-    :param float t_w: Web thickness of the PFC section
-    :param float r: Root radius of the PFC section
-    :param int n_r: Number of points discretising the root radius
-    :param shift: Vector that shifts the cross-section by *(x, y)*
-    :type shift: list[float, float]
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a PFC section with a depth of 250, a width of 90, a flange
-    thickness of 15, a web thickness of 8 and a root radius of 12, using 8 points to discretise the
-    root radius. A mesh is generated with a maximum triangular area of 5.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.channel_section(d=250, b=90, t_f=15, t_w=8, r=12, n_r=8)
-        geometry.create_mesh(mesh_sizes=[5.0])
-
-    ..  figure:: ../images/sections/pfc_geometry.png
-        :align: center
-        :scale: 75 %
-
-        PFC geometry.
-
-    ..  figure:: ../images/sections/pfc_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-
-    # add first three points
-    points.append([0, 0])
-    points.append([b, 0])
-    points.append([b, t_f])
-
-    # construct the bottom right radius
-    pt = [t_w + r, t_f + r]
-    points += draw_radius(pt, r, 1.5 * np.pi, n_r, False)
-
-    # construct the top right radius
-    pt = [t_w + r, d - t_f - r]
-    points += draw_radius(pt, r, np.pi, n_r, False)
-
-    # add last three points
-    points.append([b, d - t_f])
-    points.append([b, d])
-    points.append([0, d])
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def tapered_flange_channel(d, b, t_f, t_w, r_r, r_f, alpha, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a Tapered Flange Channel section with the bottom left corner at the origin
-    *(0, 0)*, with depth *d*, width *b*, mid-flange thickness *t_f*, web thickness *t_w*, root
-    radius *r_r*, flange radius *r_f* and flange angle *alpha*, using *n_r* points to construct the
-    radii.
-
-    :param float d: Depth of the Tapered Flange Channel section
-    :param float b: Width of the Tapered Flange Channel section
-    :param float t_f: Mid-flange thickness of the Tapered Flange Channel section (measured at the
-        point equidistant from the face of the web to the edge of the flange)
-    :param float t_w: Web thickness of the Tapered Flange Channel section
-    :param float r_r: Root radius of the Tapered Flange Channel section
-    :param float r_f: Flange radius of the Tapered Flange Channel section
-    :param float alpha: Flange angle of the Tapered Flange Channel section (degrees)
-    :param int n_r: Number of points discretising the radii
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a Tapered Flange Channel section with a depth of 10, a width of
-    3.5, a mid-flange thickness of 0.575, a web thickness of 0.475, a root radius of 0.575, a
-    flange radius of 0.4 and a flange angle of 8°, using 16 points to discretise the radii. A mesh
-    is generated with a maximum triangular area of 0.02::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.tapered_flange_channel(
-            d=10, b=3.5, t_f=0.575, t_w=0.475, r_r=0.575, r_f=0.4, alpha=8, n_r=16
-        )
-        geometry.create_mesh(mesh_sizes=[0.02])
-
-    ..  figure:: ../images/sections/taperedchannel_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Tapered flange channel geometry.
-
-    ..  figure:: ../images/sections/taperedchannel_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-
-    # calculate alpha in radians
-    alpha_rad = np.pi * alpha / 180
-
-    # calculate the height of the flange toe and dimensions of the straight
-    x1 = b * 0.5 - t_w * 0.5 - r_f * (1 - np.sin(alpha_rad))
-    y1 = x1 * np.tan(alpha_rad)
-    x2 = b * 0.5 - t_w * 0.5 - r_r * (1 - np.sin(alpha_rad))
-    y2 = x2 * np.tan(alpha_rad)
-    y_t = t_f - y1 - r_f * np.cos(alpha_rad)
-
-    # add first two points
-    points.append([0, 0])
-    points.append([b, 0])
-
-    # construct the bottom right flange toe radius
-    if r_f == 0:
-        points.append([b, y_t])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-
-            # calculate the locations of the radius points
-            x = b - r_f + r_f * np.cos(theta)
-            y = y_t + r_f * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the bottom right root radius
-    if r_r == 0:
-        points.append([t_w, t_f + y2])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = (3.0 / 2 * np.pi - alpha_rad) - (
-                i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-            )
-
-            # calculate the locations of the radius points
-            x = t_w + r_r + r_r * np.cos(theta)
-            y = t_f + y2 + r_r * np.cos(alpha_rad) + r_r * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the top right root radius
-    if r_r == 0:
-        points.append([t_w, d - t_f - y2])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = np.pi - i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-
-            # calculate the locations of the radius points
-            x = t_w + r_r + r_r * np.cos(theta)
-            y = d - t_f - y2 - r_r * np.cos(alpha_rad) + r_r * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # construct the top right flange toe radius
-    if r_f == 0:
-        points.append([b, d - y_t])
-    else:
-        for i in range(n_r):
-            # determine polar angle
-            theta = (3.0 * np.pi / 2 + alpha_rad) + (
-                i * 1.0 / max(1, n_r - 1) * (np.pi * 0.5 - alpha_rad)
-            )
-
-            # calculate the locations of the radius points
-            x = b - r_f + r_f * np.cos(theta)
-            y = d - y_t + r_f * np.sin(theta)
-
-            # append the current points to the points list
-            points.append([x, y])
-
-    # add the final two points
-    points.append([b, d])
-    points.append([0, d])
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def tee_section(d, b, t_f, t_w, r, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a Tee section with the top left corner at *(0, d)*, with depth *d*, width *b*,
-    flange thickness *t_f*, web thickness *t_w* and root radius *r*, using *n_r* points to
-    construct the root radius.
-
-    :param float d: Depth of the Tee section
-    :param float b: Width of the Tee section
-    :param float t_f: Flange thickness of the Tee section
-    :param float t_w: Web thickness of the Tee section
-    :param float r: Root radius of the Tee section
-    :param int n_r: Number of points discretising the root radius
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a Tee section with a depth of 200, a width of 100, a flange
-    thickness of 12, a web thickness of 6 and a root radius of 8, using 8 points to discretise the
-    root radius. A mesh is generated with a maximum triangular area of 3.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.tee_section(d=200, b=100, t_f=12, t_w=6, r=8, n_r=8)
-        geometry.create_mesh(mesh_sizes=[3.0])
-
-    ..  figure:: ../images/sections/tee_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Tee section geometry.
-
-    ..  figure:: ../images/sections/tee_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    points = []
-    # add first two points
-    points.append([b * 0.5 - t_w * 0.5, 0])
-    points.append([b * 0.5 + t_w * 0.5, 0])
-
-    # construct the top right radius
-    pt = [b * 0.5 + t_w * 0.5 + r, d - t_f - r]
-    points += draw_radius(pt, r, np.pi, n_r, False)
-
-    # add next four points
-    points.append([b, d - t_f])
-    points.append([b, d])
-    points.append([0, d])
-    points.append([0, d - t_f])
-
-    # construct the top left radius
-    pt = [b * 0.5 - t_w * 0.5 - r, d - t_f - r]
-    points += draw_radius(pt, r, 0.5 * np.pi, n_r, False)
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def angle_section(d, b, t, r_r, r_t, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs an angle section with the bottom left corner at the origin *(0, 0)*, with depth
-    *d*, width *b*, thickness *t*, root radius *r_r* and toe radius *r_t*, using *n_r* points to
-    construct the radii.
-
-    :param float d: Depth of the angle section
-    :param float b: Width of the angle section
-    :param float t: Thickness of the angle section
-    :param float r_r: Root radius of the angle section
-    :param float r_t: Toe radius of the angle section
-    :param int n_r: Number of points discretising the radii
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates an angle section with a depth of 150, a width of 100, a thickness
-    of 8, a root radius of 12 and a toe radius of 5, using 16 points to discretise the radii. A
-    mesh is generated with a maximum triangular area of 2.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.angle_section(d=150, b=100, t=8, r_r=12, r_t=5, n_r=16)
-        geometry.create_mesh(mesh_sizes=[2.0])
-
-    ..  figure:: ../images/sections/angle_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Angle section geometry.
-
-    ..  figure:: ../images/sections/angle_mesh.png
-        :align: center
-        :scale: 75 %
-    """
-    if r_t > t:
-        raise ValueError(
-            "The radius of the toe (r_t) cannot be larger than the toe thickness (t)."
-        )
-
-    points = []
-
-    # add first two points
-    points.append([0, 0])
-    points.append([b, 0])
-
-    # construct the bottom toe radius
-    pt = [b - r_t, t - r_t]
-    points += draw_radius(pt, r_t, 0, n_r)
-
-    # construct the root radius
-    pt = [t + r_r, t + r_r]
-    points += draw_radius(pt, r_r, 1.5 * np.pi, n_r, False)
-
-    # construct the top toe radius
-    pt = [t - r_t, d - r_t]
-    points += draw_radius(pt, r_t, 0, n_r)
-
-    # add the next point
-    points.append([0, d])
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def cee_section(d, b, l, t, r_out, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a Cee section (typical of cold-formed steel) with the bottom left corner at the
-    origin *(0, 0)*, with depth *d*, width *b*, lip *l*, thickness *t* and outer radius *r_out*,
-    using *n_r* points to construct the radius. If the outer radius is less than the thickness
-    of the Cee Section, the inner radius is set to zero.
-
-    :param float d: Depth of the Cee section
-    :param float b: Width of the Cee section
-    :param float l: Lip of the Cee section
-    :param float t: Thickness of the Cee section
-    :param float r_out: Outer radius of the Cee section
-    :param int n_r: Number of points discretising the outer radius
-    :raises Exception: Lip length must be greater than the outer radius
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a Cee section with a depth of 125, a width of 50, a lip of 30, a
-    thickness of 1.5 and an outer radius of 6, using 8 points to discretise the radius. A mesh is
-    generated with a maximum triangular area of 0.25::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.cee_section(d=125, b=50, l=30, t=1.5, r_out=6, n_r=8)
-        geometry.create_mesh(mesh_sizes=[0.25])
-
-    ..  figure:: ../images/sections/cee_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Cee section geometry.
-
-    ..  figure:: ../images/sections/cee_mesh.png
-        :align: center
-        :scale: 75 %
-    """
-    # ensure the lip length is greater than the outer radius
-    if l < r_out:
-        raise Exception("Lip length must be greater than the outer radius")
-
-    points = []
-
-    # calculate internal radius
-    r_in = max(r_out - t, 0)
-
-    # construct the outer bottom left radius
-    points += draw_radius([r_out, r_out], r_out, np.pi, n_r)
-
-    # construct the outer bottom right radius
-    points += draw_radius([b - r_out, r_out], r_out, 1.5 * np.pi, n_r)
-
-    if r_out != l:
-        # add next two points
-        points.append([b, l])
-        points.append([b - t, l])
-
-    # construct the inner bottom right radius
-    points += draw_radius([b - t - r_in, t + r_in], r_in, 0, n_r, False)
-
-    # construct the inner bottom left radius
-    points += draw_radius([t + r_in, t + r_in], r_in, 1.5 * np.pi, n_r, False)
-
-    # construct the inner top left radius
-    points += draw_radius([t + r_in, d - t - r_in], r_in, np.pi, n_r, False)
-
-    # construct the inner top right radius
-    points += draw_radius([b - t - r_in, d - t - r_in], r_in, 0.5 * np.pi, n_r, False)
-
-    if r_out != l:
-        # add next two points
-        points.append([b - t, d - l])
-        points.append([b, d - l])
-
-    # construct the outer top right radius
-    points += draw_radius([b - r_out, d - r_out], r_out, 0, n_r)
-
-    # construct the outer top left radius
-    points += draw_radius([r_out, d - r_out], r_out, 0.5 * np.pi, n_r)
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def zed_section(d, b_l, b_r, l, t, r_out, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a zed section with the bottom left corner at the origin *(0, 0)*, with depth *d*,
-    left flange width *b_l*, right flange width *b_r*, lip *l*, thickness *t* and outer radius
-    *r_out*, using *n_r* points to construct the radius. If the outer radius is less than the
-    thickness of the Zed Section, the inner radius is set to zero.
-
-    :param float d: Depth of the zed section
-    :param float b_l: Left flange width of the Zed section
-    :param float b_r: Right flange width of the Zed section
-    :param float l: Lip of the Zed section
-    :param float t: Thickness of the Zed section
-    :param float r_out: Outer radius of the Zed section
-    :param int n_r: Number of points discretising the outer radius
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a zed section with a depth of 100, a left flange width of 40, a
-    right flange width of 50, a lip of 20, a thickness of 1.2 and an outer radius of 5, using 8
-    points to discretise the radius. A mesh is generated with a maximum triangular area of 0.15::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.zed_section(d=100, b_l=40, b_r=50, l=20, t=1.2, r_out=5, n_r=8)
-        geometry.create_mesh(mesh_sizes=[0.15])
-
-    ..  figure:: ../images/sections/zed_geometry.png
-        :align: center
-        :scale: 75 %
-
-        zed section geometry.
-
-    ..  figure:: ../images/sections/zed_mesh.png
-        :align: center
-        :scale: 75 %
-    """
-    # ensure the lip length is greater than the outer radius
-    if l < r_out:
-        raise Exception("Lip length must be greater than the outer radius")
-
-    points = []
-
-    # calculate internal radius
-    r_in = max(r_out - t, 0)
-
-    # construct the outer bottom left radius
-    points += draw_radius([r_out, r_out], r_out, np.pi, n_r)
-
-    # construct the outer bottom right radius
-    points += draw_radius([b_r - r_out, r_out], r_out, 1.5 * np.pi, n_r)
-
-    if r_out != l:
-        # add next two points
-        points.append([b_r, l])
-        points.append([b_r - t, l])
-
-    # construct the inner bottom right radius
-    points += draw_radius([b_r - t - r_in, t + r_in], r_in, 0, n_r, False)
-
-    # construct the inner bottom left radius
-    points += draw_radius([t + r_in, t + r_in], r_in, 1.5 * np.pi, n_r, False)
-
-    # construct the outer top right radius
-    points += draw_radius([t - r_out, d - r_out], r_out, 0, n_r)
-
-    # construct the outer top left radius
-    points += draw_radius([t - b_l + r_out, d - r_out], r_out, 0.5 * np.pi, n_r)
-
-    if r_out != l:
-        # add the next two points
-        points.append([t - b_l, d - l])
-        points.append([t - b_l + t, d - l])
-
-    # construct the inner top left radius
-    points += draw_radius([2 * t - b_l + r_in, d - t - r_in], r_in, np.pi, n_r, False)
-
-    # construct the inner top right radius
-    points += draw_radius([-r_in, d - t - r_in], r_in, 0.5 * np.pi, n_r, False)
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-zee_section = zed_section  # An alias for our American friends (and friends who use "American English")
-
-
-def cruciform_section(d, b, t, r, n_r, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a cruciform section centered at the origin *(0, 0)*, with depth *d*, width *b*,
-    thickness *t* and root radius *r*, using *n_r* points to construct the root radius.
-
-    :param float d: Depth of the cruciform section
-    :param float b: Width of the cruciform section
-    :param float t: Thickness of the cruciform section
-    :param float r: Root radius of the cruciform section
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-
-    The following example creates a cruciform section with a depth of 250, a width of 175, a
-    thickness of 12 and a root radius of 16, using 16 points to discretise the radius. A mesh is
-    generated with a maximum triangular area of 5.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.cruciform_section(d=250, b=175, t=12, r=16, n_r=16)
-        geometry.create_mesh(mesh_sizes=[5.0])
-
-    ..  figure:: ../images/sections/cruciform_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Cruciform section geometry.
-
-    ..  figure:: ../images/sections/cruciform_mesh.png
-        :align: center
-        :scale: 75 %
-    """
-    points = []
-
-    # add first two points
-    points.append([-t * 0.5, -d * 0.5])
-    points.append([t * 0.5, -d * 0.5])
-
-    # construct the bottom right radius
-    pt = [0.5 * t + r, -0.5 * t - r]
-    points += draw_radius(pt, r, np.pi, n_r, False)
-
-    # add the next two points
-    points.append([0.5 * b, -t * 0.5])
-    points.append([0.5 * b, t * 0.5])
-
-    # construct the top right radius
-    pt = [0.5 * t + r, 0.5 * t + r]
-    points += draw_radius(pt, r, 1.5 * np.pi, n_r, False)
-
-    # add the next two points
-    points.append([t * 0.5, 0.5 * d])
-    points.append([-t * 0.5, 0.5 * d])
-
-    # construct the top left radius
-    pt = [-0.5 * t - r, 0.5 * t + r]
-    points += draw_radius(pt, r, 0, n_r, False)
-
-    # add the next two points
-    points.append([-0.5 * b, t * 0.5])
-    points.append([-0.5 * b, -t * 0.5])
-
-    # construct the bottom left radius
-    pt = [-0.5 * t - r, -0.5 * t - r]
-    points += draw_radius(pt, r, 0.5 * np.pi, n_r, False)
-
-    polygon = Polygon(points)
-    return Geometry(polygon, material)
-
-
-def polygon_hollow_section(d, t, n_sides, r_in=0, n_r=1, rot=0, material: pre.Material = pre.DEFAULT_MATERIAL) -> Geometry:
-    """Constructs a regular hollow polygon section centered at *(0, 0)*, with a pitch circle
-    diameter of bounding polygon *d*, thickness *t*, number of sides *n_sides* and an optional
-    inner radius *r_in*, using *n_r* points to construct the inner and outer radii (if radii is
-    specified).
-
-    :param float d: Pitch circle diameter of the outer bounding polygon (i.e. diameter of circle
-        that passes through all vertices of the outer polygon)
-    :param float t: Thickness of the polygon section wall
-    :param float r_in: Inner radius of the polygon corners. By default, if not specified, a polygon
-        with no corner radii is generated.
-    :param int n_r: Number of points discretising the inner and outer radii, ignored if no inner
-        radii is specified
-    :param rot: Initial counterclockwise rotation in degrees. By default bottom face is aligned
-        with x axis.
-    :param Optional[sectionproperties.pre.pre.Material]: Material to associate with this geometry
-    :raises Exception: Number of sides in polygon must be greater than or equal to 3
-
-    The following example creates an Octagonal section (8 sides) with a diameter of 200, a
-    thickness of 6 and an inner radius of 20, using 12 points to discretise the inner and outer
-    radii. A mesh is generated with a maximum triangular area of 5::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.polygon_hollow_section(d=200, t=6, n_sides=8, r_in=20, n_r=12)
-        geometry.create_mesh(mesh_sizes=[5])
-
-    ..  figure:: ../images/sections/polygon_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Octagonal section geometry.
-
-    ..  figure:: ../images/sections/polygon_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    outer_points = []
-    inner_points = []
-
-    if n_sides < 3:
-        msg = "n_sides required to be greater than 3 for polygon_section()"
-        raise Exception(msg)
-
-    # initial rotation
-    rot = rot * np.pi / 180  # radians
-
-    # determine triangular segment angle
-    alpha = 2 * np.pi / n_sides  # radians
-
-    # determine distance from origin to point perpendicular on face of side
-    a_out = d / 2 * np.cos(alpha / 2)
-    a_in = a_out - t
-
-    # determine side length for outer & inner faces neglecting radii
-    side_length_out = d * np.sin(alpha / 2)
-    side_length_in = a_in / a_out * side_length_out
-
-    # check limit on internal radii, if exceeded then radii merge to circle
-    if r_in > a_in:
-        r_in = a_in
-        circle = True
-    else:
-        circle = False
-
-    # calculate external radius, if r_in is zero, r_out also is zero
-    if r_in == 0:
-        r_out = 0
-        n_r = 1
-    else:
-        r_out = r_in + t
-
-    # equivalent side length of half the corner radii triangular segment
-    c_out = r_out * (side_length_out / 2) / a_out
-    c_in = r_in * (side_length_in / 2) / a_in
-
-    # determine straight side length between corner radii (if present)
-    side_length_straight_out = side_length_out - (2 * c_out)
-    side_length_straight_in = side_length_in - (2 * c_in)
-
-    # temp list for repeating geometry
-    outer_base_points = []
-    inner_base_points = []
-
-    # start at bottom face, constructing one corner radii, then rotate by initial rotation +
-    # alpha and repeat for n_side number of times to form full section perimeter
-
-    # construct the first radius (bottom right)
-    for i in range(n_r):
-        # determine polar angle
-        theta = 1 / 2 * np.pi + i * 1.0 / max(1, n_r - 1) * alpha
-
-        # calculate location of inner and outer points
-        x_outer = side_length_straight_out / 2 - r_out * np.cos(theta)
-        y_outer = -a_out + r_out - r_out * np.sin(theta)
-        x_inner = side_length_straight_in / 2 - r_in * np.cos(theta)
-        y_inner = -a_in + r_in - r_in * np.sin(theta)
-
-        # append the current temporary points to the temporary points list
-        outer_base_points.append([x_outer, y_outer])
-        inner_base_points.append([x_inner, y_inner])
-
-    for i in range(n_sides):
-        for point in outer_base_points:
-            point_new = rotate(point, alpha * i + rot)
-            outer_points.append(point_new)
-
-        for point in inner_base_points:
-            point_new = rotate(point, alpha * i + rot)
-            inner_points.append(point_new)
-
-    outer_polygon = Polygon(outer_points)
-    inner_polygon = Polygon(inner_points)
-    return Geometry(outer_polygon - inner_polygon, material)
-
-
-def rotate(point, angle: float):
-    """
-    Rotate a point counterclockwise by a given angle around origin [0, 0]
-
-    :param list point: Point coordinates to be rotated
-    :param float angle: Angle to rotate point coordinates
-    :return: Coordinates of rotated point
-    :rtype: list[float, float]
-    """
-
-    pt_x, pt_y = point
-
-    c = np.cos(angle)
-    s = np.sin(angle)
-
-    new_x = c * pt_x - s * pt_y
-    new_y = s * pt_x + c * pt_y
-
-    return [new_x, new_y]
-
-
-def box_girder_section(d, b_t, b_b, t_ft, t_fb, t_w, material: pre.Material = pre.DEFAULT_MATERIAL):
-    """Constructs a box girder section centered at at *(max(b_t, b_b)/2, d/2)*, with depth *d*, top
-    width *b_t*, bottom width *b_b*, top flange thickness *t_ft*, bottom flange thickness *t_fb*
-    and web thickness *t_w*.
-
-    :param float d: Depth of the Box Girder section
-    :param float b_t: Top width of the Box Girder section
-    :param float b_b: Bottom width of the Box Girder section
-    :param float t_ft: Top flange thickness of the Box Girder section
-    :param float t_fb: Bottom flange thickness of the Box Girder section
-    :param float t_w: Web thickness of the Box Girder section
-
-    The following example creates a Box Girder section with a depth of 1200, a top width of 1200, a
-    bottom width of 400, a top flange thickness of 16, a bottom flange thickness of 12 and a web
-    thickness of 8. A mesh is generated with a maximum triangular area of 5.0::
-
-        import sectionproperties.pre.sections as sections
-
-        geometry = sections.box_girder_section(d=1200, b_t=1200, b_b=400, t_ft=100, t_fb=80, t_w=50)
-        geometry.create_mesh(mesh_sizes=[200.0])
-
-    ..  figure:: ../images/sections/box_girder_geometry.png
-        :align: center
-        :scale: 75 %
-
-        Box Girder geometry.
-
-    ..  figure:: ../images/sections/box_girder_mesh.png
-        :align: center
-        :scale: 75 %
-
-        Mesh generated from the above geometry.
-    """
-    outer_points = []
-    inner_points = []
-
-    # calculate central axis
-    x_c = max(b_t, b_b) * 0.5
-
-    # determine side wall angle
-    if b_t < b_b:
-        phi_b = np.arctan2(d, 0.5 * (b_b - b_t))
-        phi_t = np.pi - phi_b
-    else:
-        phi_t = np.arctan2(d, 0.5 * (b_t - b_b))
-        phi_b = np.pi - phi_t
-
-    # determine inner wall x-offsets
-    x_bot = t_fb / np.tan(np.pi - phi_b)
-    x_top = t_ft / np.tan(np.pi - phi_t)
-    web_x = abs(t_w / np.sin(np.pi - phi_b))
-
-    # add outer points
-    outer_points.append([x_c - 0.5 * b_b, 0])
-    outer_points.append([x_c + 0.5 * b_b, 0])
-    outer_points.append([x_c + 0.5 * b_t, d])
-    outer_points.append([x_c - 0.5 * b_t, d])
-
-    # add inner points
-    inner_points.append([x_c - 0.5 * b_b - x_bot + web_x, t_fb])
-    inner_points.append([x_c + 0.5 * b_b + x_bot - web_x, t_fb])
-    inner_points.append([x_c + 0.5 * b_t + x_top - web_x, d - t_ft])
-    inner_points.append([x_c - 0.5 * b_t - x_top + web_x, d - t_ft])
-
-    outer_polygon = Polygon(outer_points)
-    inner_polygon = Polygon(inner_points)
-
-    return Geometry(outer_polygon - inner_polygon, material)
